Remove checking tsc:deps from lint script
On a fresh install of the repo, the tarballs for dependencies such as preact-local won't be present and so those dependencies won't have preact locally installed. Therefore typechecking will fail in those dependencies. Our CI is a fresh install of the repository and so the CI will always fail this typecheck.

For the time being until a better workaround is found, I'm removing typechecking deps from the CI